Remove incorrect parent for draggable ship copy set to body
Identified by comparing any discrepancies between the original code
snippet and by adapted code after seeing a relatively positioned parent
for the relatively positioned div still allowed it to move. Here the
draggable copy is in the wrong parent even if the original ship picture
is in the correct parent in the board, even though its parent does not
matter.