Give better initial starting position to copy ship before moved
This is not present in the original snippet code so it is a judgement
call. Place it a position that is the same as the original ship. The
current position might have been too far offset by a window scroll
position if there ever was any. There wasn't any I believe anyway. But
still. Unwanted behaviour and I am not sure where I got it from. It is
inappropriate to add it here before it is even moved.