Fix: Image does not show due to local url not mapping to hosted url
I easily fix this. By importing the img as local url it will be
converted to the hosted url. So you can just use the img from then on
and it will be with the hosted url. I saw this in the pictures_display
file working, so figured I must be able to dynamically import any images
through the local url to then use them. I don't have to manually try
resolve the hosted url for myself.