Fix: Ship not appearing after mouse down
Setting a breakpoint to the start of the dropShip method before the
check for if a ship was selected. This was revealed it was set to false
somewhere. This was in the document event listener affecting what
happened in the other listener. First it set the draggableShipType to
false in an effort to delete and set it to false. If this event
listener played before the one for the cells, these 2 things would
happen. Either the ship wouldn't place and even if it did it would be
with any empty pic url that was deleted.

For some reason going into debugger mode got the correct order and did
not allow the document event listener to dispatch first. The event order
might have been set to different priorities.