Fix: Ship never touches edge of board (misses by 1) and other glitches
There were other unspecified visual glitches with ship placement on
board. They seem all fixed. Check for changes diff in code between this
commit and the commit before the last one for a look into the real
changes. The only behavior I recall modifying is where row and col
were updated even when only one needed a shift and the other one could
stay. To that end though I needed to update row start and col start
to reflect the shift offset and not anytime before that.