Fix: SVG are not stretched out and deformed as they should to fill space
This was because aspect ratio needed to be set to not be preserved. This
idea came to me watching the video linked below at 14:13. The
characteristics of the result from not deforming aspect ratio and it
being artificially limited to not reach the edge of the viewport
reminded me of my case possibly. What happened in the video was that he
reduced the size of the viewport. Of course the viewbox was the same
size but bigger than the viewport so it meant the viewbox showed a more
zoomed in view of a now reduced viewport. Video:
[What are SVG viewport and viewBox? [ A beginners guide to SVG part 3 ]](https://www.youtube.com/watch?v=TBYJ2V1jAlA&t=3s)