Feat/Refactor: By default use player1 board for drag/drop controller
This is not the generic board it was previously set to. This is the
board tied to player 1 that is set as not the CPU board. Initially the
drag and drop is set to work on player 1 until it changes through
getters and setters to perform the drag and drop for the player 2 board,
both logical and UI board.