Feat: Selectively show ship image of desired orientation on board
The default displayed is the horizontal. But any can be retrieved from
memory depending on the circumstance.

It does not have to be "on" anywhere physically to use it as a
reference. It must just be a reference to the SVG available in memory.
We do not need to make it selectively appear until we need it.