Fix: Randomizing vertical ships leads to infinite or over 5 ships
Variable `i` does NOTHING on overlapping ship placement detection with
return. Don't move it backward and decrement it. Just keep it stuck.
That was what I did different with the horizontal. Fixed. That was the
mistaken key. I did not even realize it. In fact I accidentally forgot
the return of -1 for horizontal overlap result. Therefore the
decrementing of i taking it back to overcompensate for both the current
ship placement with both putting an extra previous ship of i iteration
and the ship of the actual iteration you were busy on.

And in moving back and placing too much it might have filled the board
too much and it would loop endlessly trying to fulfil i to the end. For
some reason another there was another bug where I could visualize placed
horizontal ships with each iteration but not vertical ones, until the
loop was complete. This seems like an bug not related to the current.
So it makes it difficult to confirm this as the cause of the infinite
loop. Inspection of DOM elements is not possible with page freeze.