Fix: Module returns empty object on import from circular reference
This was due to a circular reference between gameManager and
setupManager. The former imports the latter and when it gets to that
import to try execute the file to therefore import its functionality it
finds that file needs to import its own functionality first to start
functioning.

You can break the circular reference by injecting the dependency in
other ways. I did it with a circumstantial require which is not a neat
solution and seems a bit arbitrary I'll admit.