Refactor: Try decouple logical placement from UI placement
To call them separately in case of wanting to place overlapping ships
in UI even if they are still placed in logical board. This is in order
to technically place ships twice without overlapping being detected.
This is for when placing ships when repopulating the UI board to not
have to have the logical placement be replaced or worry about checking
overlap, only UI board needs to check that visually, first and logical
placement will check that after.