Feat: Add cell wrapper objects to later replace ships on logical board
Cell wrappers will contain a ship object or none if that cell in the
logical board has no ship. These are added for extra functionality of
knowing which space in the logical board is marked as a hit. A ship
object is cohesive across all the grid slots it occupies. There are
records of misses on the logical board. There are no records of hits.
If we place a hit marker we'd have to replace the ship. So rather we
contain all this information inside unique cells which are not cohesive
copies like the ship.

Information of a hit is important to allow replays if the player double
clicks on a cell they already clicked and to not pass control to CPU.
This happens when detecting misses not hits on ships. We don't know what
spot the ship was hit if it was hit twice.