Fix: Cross marker appended to player board after game ends
Control was passed to CPU which played on player 1 board, the final play
which is what the CPU should not be doing. Control should not pass if
ships have sunk. Bug found by stepping through end game scenario in
debugger to see how the construction the view happened and where else
the problem was introduced.