Fix: Prevent reset immediately after trying down move even if success
Reset only must happen if last move is down and it produces a failure.
Only then should you reset everything.

As it went on by some chance and went on to next move alternated (5)
which is non-existent so reset. I reset all the time which is a problem.
Furthermore I fail to reset on failure and I continue move alternator
pos from where it left off. This is before the down move. Like if we
move on to a next random ship with success after we sunk the last
ship sucessfully but never reset the moves to try. We will fix this
next.