Fix: Stop order of next move being left, right, down, up, back to down
It always went down instead of up then it goes back to doing up move
properly. So what was happening, prematurely going up?
This is the pattern whether or not we start with the vertically oriented
ship or it comes up later during random striking. It should try up
before trying down. It always goes down then up then back down.
It did not happen with other directions. So I checked the unique problem
with having to go down and the next possible moves was nudged up instead
of down. And the next up move was down instead of up. Because of how
I change the y value.