Fix: Allow hits to be regarded as marked spots to skip not just misses
If an X marked a spot. And it was being considered as a next move it
should be skipped just like if there was a miss in that area. For some
reason I only considered misses. Ships that were hit must also be
skipped when finding new moves, so don't just look beyond them, in the
same direction.