Fix: Stop wrong value being used on skip to decide next move
The original value is used. The next value is for the move it was
supposed to be. The value doesn't reflect the value it truly is in the
part of the array it truly is. That is why the next move value is simply
wrong. It adopts the value it was meant to be for the previous move not
the current move beforehand if it skipped one move to go to another.